Add server
Added the server main namespace. This will have a `-main` fn that starts
aleph HTTP server and can handle compojure routing. Only added a dummy
controller for the index page.

Also worked on the project.clj file and started a profile just for clj
part of the project. Added an alias:
* `lein server` to start the server
* `lein test` that will run all the tests