Add travis for CI
Travis CI will run all branches when updated. I already added a test so
it should be already doing something.

Also added the badge to README and improved it a bit.

Had to add a custom `script/ci` that fetches latest lein, because the
version on travis is older and I could not find a way to upgrade it. So
we get the lein script and it installs. Didn't notice a slowdown.