Refactor controller utilities
Added a reply namespace that receives an object and returns an HTTP map
witht the object as edn. Also added a mock-request namespace to test
controllers. For now, it only has an http-get fn that returns the body
as a clojure object (reads the edn string).

Also added the result lib to handle result-based operations.