Boot clojurescript and figwheel
Added a new cljs-dev profile that is ready for figwheel development.
Created an `lein frontend` alias that starts figwheel and starts the
frontend server at `http://localhost:3450`.

Added a dev.html page, that will be the entry point for development.
Usually the prd endpoint is different, has optimized JS, analytics and
other dependencies. The dev.html will be the starting point for
development.