Added dev.html entry page
Bootstrapped the html dev page and, on page load, made it perform a REST
request to the root of the server, that returns server info. It then
prints to the browser console the response, in edn.

There is still room for improvement, for example we may need an
indirection to handle the edn deserialization and overal requests. But
I'll leave that for later.