Start events module
Motivation:

We need an event module to handle events on the game, mainly game
changed events and lobby events.

Modifications:

Added a new models module with the objective of having state/property
bags that represent the game models/entities. It should be available on
the server and on the client. Added a `game-info` model. Didn't use
`game` because that's a different concept on the rules engine. Game info
will have the game from roles, player identifiers and specific state.

Created a AsyncPubSub component that will be responsible for
implementing generic pub/sub logic on top of core.async.

Created a GamesEventSystem that will be responsible for managing the
game events. It dependes on a `:pub-sub-impl` that at this moment is
AsyncPubSub. It doesn't do anything yet.

This commit is just the skeleton of what's to come.

Issue #3
PR #4