Implement pub/sub for game events
Motivation:

Game events needs to be implemented and actually do something.

Modifications:

Had to add a generic PubSub protocol to interface the implementation of
the pub sub via core.async. This protocol is now used on the
GameEventsSystem.

Had to learn about async/pub, async/sub, async/mult and async/tab to
implement AsyncPubSub. When this component starts, creates an channel
named `master-ch` that will be the input channel to send events to.

Then, we create a mult for it, named broadcast. Now all events sent to
`master-ch` will be sent to channels that were tapped on the broadcast.

Next we tap a `publisher-ch` to the `master-ch`. This is to support a
subscribe on a specific event name.

Also added a unit test for this flow. Still need to do the unsubscribe.

Issue #3
PR #4