Allow subscribe to game
Motivation:

We need to be able to subscribe to a specific game and get only its
events.

Modifications:

Started by adding an id to the game-info. At this moment it's a random
value.

Then I realized that I would not be able to have a shared core
async impl for the game-events and for the future lobby-events, because
the lobby doesn't need the specific publications that game-events have.

So I removed that files/namespaces and moved the implementation directly
to GameEventsSystem.

After that I started getting deadlocks on the auto test runner. It would
happen only after 5-15 runs and ramdonly. After messing around with
core.async code I found the issue. I was creating one `publisher-ch` and
reusing it on both events and games publications. But how the pub works,
is by having a go-loop that extracts the value from the channel and
broadcasts it to the subscribers.

Because both pubs were using the same publisher channel, sometimes one
of them would pakk when it should have something, because the other had
already removed it. On the tests, if we blocked on this scenario, we
would deadlock.

So I created two publisher channels and tapped them with the broadcast
mult, and the issue was gone.

Issue #3
PR #4