Add lobby events component
Motivation:

We need a component responsible for handling lobby events. There are two
main events: when a game is added to the lobby, and when a game is
removed to the lobby. Subscribers will be able to get this events, and
by so be notified everytime the lobby changes.

Modifications:

Added the lobby-events ns, with the LobbyEvents protocol, and the
LobbyEventsSystem component implementation, that relies on core.async.

Ref: #3