Transform http server to a component and some refactoring
Motivation:

Create a component interface for the webserver, to have it on the system
and to start it and stop it cleanly.

Modifications:

Created the system namespace that creates a system. Also created a
test-system namespace with a system just for tests. At this time is
basically the original system, without the :http-component.

Then I bumped into a problem. The handlers need the full system, but I
don't have access to it. When the components are started, they can
specify specific dependencies, but I actually want the full system.

I searched for patterns to handle this and didn't find anything useful.
Sometimes handlers are implemented as components, and thus can specify
dependecies, but I think that is too much.

So my solution was to create the system, and auto assoc it on a :mega
key, as a map (this way it won't be started again).