Fix lein run and shutdown hook
Motivation:

Running `lein run` was starting the system and then exiting right away.
Shutdown hook only ran sometimes when issuing ctrl+c.

Modifications:

The server problem is that aleph starts daemon threads, and so the jvm
doesn't detect any long lived thread and shuts down. To provent this I
added a channel to the http server component named closed-ch where the
main thread can block on.

The ctrl+c issue was related to lein. Lein starts a nested jvm to run
the app, and was lein and not the app that was cathing the ctrl+c. Now I
added the trampoline argument that makes lein run the app on the master
jvm, and this way the hook is run.