Add create-lobby-game interactor
Motivation:

We want an interactor to create and add a new game to the lobby.

Modifications:

Added a new interactor, that receives the system and args/params to
create the game. At this moment it will create a dummy game-info and set
the p1 id. After that it will broadcast the new created game.

The interactor's run! fn returns a channel that will have eventually a
result for the operation.