Configure Leiningen to access private repositories (for Datomic)
Context:
The build on Travis CI was failing because Leiningen was unable to
download Datomic. This dependency must be downloaded from a private
repository, so we need to set up the Leiningen authentication to it.

In order for the build to work, the changes made here are not enough.
It was also needed to set the environment variables
`DATOMIC_REPO_USERNAME` and `DATOMIC_REPO_PASSWORD` on the settings
section of the Travis web application.

Follow [this link](https://github.com/technomancy/leiningen/blob/master/doc/DEPLOY.md#authentication)
to learn about this subject.

Motivation:
Being able to fetch dependencies like Datomic from private repositories.

Changes:
* Add information about the Datomic private repo to `project.clj`;